Store explicit TaskList partition data
Replace num_read_partitions and num_write_partitions with an explicit map of partition ids to partition configuration. This enables assigning isolation groups to partitions in the future.

This change is backwards compatible as it populates both values when writing data to Cassandra and when returning it via the API. When draining partitions we continue to maintain a contiguous block of partition ids.