Use tackAckManager's BacklogCount for DescribeTaskList.TaskListStatus.BacklogCountHint
In all other contexts we use the taskAckManager's BacklogCount as the BacklogCountHint, such as the LoadBalancerHints and task poll responses. This definition aligns with those and removes the error-prone logic to use another value, silently falling back to this value.

This additionally ensures that DescribeTaskList is responsive as all values returned are stored in memory. This method is used by the UI, CLI, and in the future will be used for automated tasklist partition scaling.