Implement TaskListStatus.empty and DescribeTaskListReponse.TaskList
Return the TaskList as part of DescribeTaskListReponse so that the TaskListKind is returned to the client. TaskListKind isn't considered part of TaskList identity, so while it's specified as part of the request it doesn't actually mean anything unless it's the request that is creating the TaskList.

Return an explicit indicator that a TaskList has no tasks within it. A subsequent change will update the adaptive scaler to use this new field.