Actually-minimal histogram impl, just extend our existing scopes
this also customizes the linter to find duplicate metrics names via the indexes, and WOW.
almost 1000 of them.
most look reasonable / they are very likely on the same scope... but not all.  nowhere near all.

so I think this might be the best argument in favor of the struct approach, actually.