fix(active-active): Fix active workflow check for history APIs requiring runID (#7341)
<!-- Describe what has changed in this PR -->
**What changed?**
Use the active cluster info of workflow's cluster attribute to check if
workflow is active in the current cluster

<!-- Tell your future self why have you made these changes -->
**Why?**
The active cluster of a workflow should be determined by the cluster
attribute

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: Mitch O'Dea <mitchelo@netapp.com>