feat: [Shard-Distributor] Implementation of onboarding logic (#7321)
<!-- Describe what has changed in this PR -->
**What changed?**

Logic implemented in the shard distributor:
when receiving the heartbeat
- Error if the namespace is in local passthrough mode, we expect no
external calls
- local passhtrough shadow and distributed passthrough have the same
behavior: check the shard assignment for the executor, if nothing
changed return it back, if it changed then delete the executor and add
the shards again. In case these are the modality the namespace
reconciliation loop is not running for the namespace (this way we do not
reassign shards while we delete the executor)
-  no changes to the normal flow in case the mode is onboarded

Logic that will be implemented in a followup pr for executor library:
- create the module to instantiate the executor with local passthrough
or communication with SD
- the modality will be assigned after the first hearbeat request (except
if it is local passthrough which is statically assigned)
- local passhtrough shadow, check the answer against the current request
and do not assign back to the internal state
- distributed passthrough before putting into place the new sharding
assignment, send heartbeat and applied it after receiving it back
- onboarded normal flow

NextNext PR
For each of the cases create a test in the canary

<!-- Tell your future self why have you made these changes -->
**Why?**

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests and local execution

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
the only service using it, it's the canary. No real impact in production

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>
Signed-off-by: Mitch O'Dea <mitchelo@netapp.com>